refactor(core): deepen Candidate construction
