+++ .changeset/deepen-candidate-construction.md
@@ -1,0 +1,7 @@
+---
+"@ggsvelte/core": patch
+---
+
+# Deepen Candidate construction
+
+Consolidate source-backed and identity-indexed Candidate construction behind one pipeline module seam while preserving CandidateStore behavior and performance contracts.
